get all new beers & refact
